Add a test for EventActivityIdControl along with some small modifications to the API itself.

+++ src/mscorlib/src/System/Diagnostics/Eventing/EventPipe.cs
@@ -169,7 +169,7 @@ namespace System.Diagnostics.Tracing
         internal static extern void DeleteProvider(IntPtr provHandle);
 
         [DllImport(JitHelpers.QCall, CharSet = CharSet.Unicode)]
-        internal static extern ulong EventActivityIdControl(uint controlCode, ref Guid activityId);
+        internal static extern int EventActivityIdControl(uint controlCode, ref Guid activityId);
 
         [DllImport(JitHelpers.QCall, CharSet = CharSet.Unicode)]
         internal static extern unsafe void WriteEvent(IntPtr eventHandle, uint eventID, void* pData, uint length, Guid* activityId, Guid* relatedActivityId);
